requirements.txt
pyserial>=3.5
PyQt5>=5.15.0
